# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Absensi",
        "description": "API Aplikasi absensi berbasis web (e-absensi) online untuk memantau kinerja karyawan jarak jauh, dan pekerja work from home (WFH).",
        "termsOfService": "http://127.0.0.1:8000/terms/",
        "contact": {
            "email": "bandungcoders@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "description": "",
                "operationId": "password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "description": "Old password user",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "New password",
                                        "type": "string"
                                    },
                                    "new_confirm_password": {
                                        "description": "password confirm",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/post/delete/{id}": {
            "delete": {
                "tags": [
                    "post"
                ],
                "summary": "Deletes a post",
                "description": "",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all data users",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/all": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all data users",
                "description": "",
                "operationId": "all",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/employee": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all data of employee",
                "description": "",
                "operationId": "employee",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "bearer token"
                        ]
                    }
                ]
            }
        },
        "/user/admin": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all data of admin",
                "description": "",
                "operationId": "admin",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "bearer token"
                        ]
                    }
                ]
            }
        },
        "/user/create": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add a new user to the store",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email user",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "description": "Phone number user",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Role of user (admin or employee)",
                                        "type": "int"
                                    },
                                    "address": {
                                        "description": "Address user",
                                        "type": "string"
                                    },
                                    "position": {
                                        "description": "Position of user (Ex: Operator Produksi)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/detail/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get detail of user",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id to show detail",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/update/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email user",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Phone number user",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Role of user {admin or employee}",
                                        "type": "int32"
                                    },
                                    "address": {
                                        "description": "Address user",
                                        "type": "string"
                                    },
                                    "position": {
                                        "description": "Position of user (Ex: Operator Produksi)",
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "description": "Avatar of user",
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deletes a user by id",
                "description": "",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}